streamline stack parsing a bit - allow online lookups
