+++ src/Scripts/ui/getHeaders/GetHeaders.ts
@@ -63,6 +63,7 @@ export class GetHeaders {
     }
 
     public static async send(headersLoadedCallback: (_headers: string, apiUsed: string) => void) {
+        Errors.log(null, "GetHeaders.send", true);
         if (!GetHeaders.validItem()) {
             ParentFrame.showError(null, "No item selected", true);
             return;
